fix(vizia): font typeface, clipboard, knob drag, shared-ui button unification
Root-cause fixes affecting all 6 plugins:
- vg::Font::default() had no typeface attached, draw_str() silently drew
  nothing with it - this was the cause of every "dB numbers missing"
  report (peak meters, goniometer correlation, EQ grid labels). Now loads
  a real system typeface once via FontMgr, cached in shared-ui/canvas.rs.
- vizia's clipboard feature was never enabled (default-features=false,
  features=["baseview"] only) - paste never worked anywhere. Added
  "clipboard" feature to all 6 plugin Cargo.toml, pulls in copypasta.
- KnobView/HSliderView drag could get stuck if a MouseUp was missed
  (e.g. host stealing focus mid-drag) - dragging stayed true forever,
  ignoring input until a fresh click. Now checks the actual button state
  on every MouseMove and self-heals.

shared-ui unification (was fragmented per plugin, DeepSeek flagged it):
- New shared-ui/src/buttons.rs: BUTTON_HEIGHT/_SMALL, KNOB_SIZE,
  SLIDER_HEIGHT constants + AMBER/IDLE_BG/DANGER_BG/DANGER_TEXT colors +
  toggle_button()/toggle_button_small()/danger_button() helpers - the old
  Iced toggle_button/output_tools_strip pattern that never got ported to
  Vizia, root cause of several inconsistent/dead toggle buttons.
- Aether, Equilibrium, Meridian, Lucent migrated onto the shared helpers
  (Bypass, Mono/Delta/Solo/Listen/Pre-Master, EQ band-type, slope
  selectors, Split/Clip, Crossfeed realism, Standalone/Relay/Hybrid,
  Resonance/relay toggles, all RESET buttons). Aurum intentionally left
  on its own local kn()/tog()/cyc() - UI still WIP.

Per-plugin fixes:
- Aether: SAVE silently no-op'd with no vault path configured - now
  falls back to the plugin data dir like Meridian/Equilibrium. Vault
  setup screen rebuilt to match Lucent's card layout. EQ curve fill
  removed (shared SpectrumView eq_curve rendering). Textbox/ComboBox
  sizes reduced.
- Equilibrium: goniometer correlation dot colors restored to the
  original Iced values (had drifted during the port). Footer analyze
  section and preset list alignment fixed. Toggle/reset button heights
  reduced.
- Meridian: reset button only touched DSP, not the UI widgets - top
  strip and footer knob sections now rebuild on params_gen like
  Equilibrium already did. Width param's right-click reset used the
  wrong default (0% instead of 100%). Window height increased for
  header/footer breathing room. Peak meter extended toward the
  goniometer.
- Lucent: Sensitivity knob was fed a raw 0-100 value where KnobView
  expects a normalized 0-1 value_norm, spun into a garbled sunburst.
  Masking toggle never reacted to AnalyzeMode changes after the initial
  build. Mode switch button recolored amber. Duplicate "LX AUDIOLABS"
  sidebar label removed. Relay bar left-aligned.
- Lucent-Relay: window was 260x160, too small for the target ComboBox's
  popup to render without clipping (vizia's Popup has should_reposition
  disabled) - grown to 260x240. Textbox/ComboBox sizes reduced.

+++ plugins/equilibrium/src/lib.rs
@@ -429,25 +429,25 @@ impl PluginLogic for Equilibrium {
             let mut bands_r = [band1_r, band2_r, band3_r, band4_r, band5_r];
 
             let band_gains = [
-                db_to_gain(self.params.low_gain.value() as f32),
-                db_to_gain(self.params.bass_gain.value() as f32),
-                db_to_gain(self.params.mid_gain.value() as f32),
-                db_to_gain(self.params.high_mid_gain.value() as f32),
-                db_to_gain(self.params.high_gain.value() as f32),
+                db_to_gain(self.params.low_gain.value()),
+                db_to_gain(self.params.bass_gain.value()),
+                db_to_gain(self.params.mid_gain.value()),
+                db_to_gain(self.params.high_mid_gain.value()),
+                db_to_gain(self.params.high_gain.value()),
             ];
             let band_widths = [
-                self.params.low_width.value() as f32 / 100.0,
-                self.params.bass_width.value() as f32 / 100.0,
-                self.params.mid_width.value() as f32 / 100.0,
-                self.params.high_mid_width.value() as f32 / 100.0,
-                self.params.high_width.value() as f32 / 100.0,
+                self.params.low_width.value() / 100.0,
+                self.params.bass_width.value() / 100.0,
+                self.params.mid_width.value() / 100.0,
+                self.params.high_mid_width.value() / 100.0,
+                self.params.high_width.value() / 100.0,
             ];
             let band_pans = [
-                self.params.low_pan.value() as f32,
-                self.params.bass_pan.value() as f32,
-                self.params.mid_pan.value() as f32,
-                self.params.high_mid_pan.value() as f32,
-                self.params.high_pan.value() as f32,
+                self.params.low_pan.value(),
+                self.params.bass_pan.value(),
+                self.params.mid_pan.value(),
+                self.params.high_mid_pan.value(),
+                self.params.high_pan.value(),
             ];
             let band_solos = [s_low, s_bass, s_mid, s_high_mid, s_high];
 
@@ -529,7 +529,7 @@ impl PluginLogic for Equilibrium {
                 processed_r = out_r - cut_r;
             }
 
-            let out_gain = db_to_gain(self.params.output_gain.value() as f32);
+            let out_gain = db_to_gain(self.params.output_gain.value());
             processed_l *= out_gain;
             processed_r *= out_gain;
